fix(library): media focus-restore fix-round minors (task-31631/31634/31567)
Final whole-branch review found 1 Important (docs) + 6 Minor. This commit
addresses the code/test minors: M-1 comments that active_effect_duration=0
covering browse-mode rows too is intended (not scope creep); M-2 dedupes
the three painted-screen test helpers (_top_border_row/_painted_cells/
_row_is_painted_focused, previously duplicated across render_fixes.py and
reader_flow.py with divergent signatures) into test_library_shell.py, the
one shared module neither side already imports from in a way that would
create a cycle; M-4 adds a right-edge click to the whole-row-toggle pin;
M-5 tightens an inequality focus-move-count assertion to the exact
observed list. M-3 is declined per the review (no fix needed now). M-6
files backlog task-31636 (canvas_sync's queue_after_recompose(None)
failure paths clearing a caller's queued follow-up, all four canvas
kinds) and task-31637 (sibling canvases still swallow a fast second click
the same way media did before Task 2's fix) so both accepted riders have
a tracked home instead of living only in the SDD ledger.

All 7 named test files re-run in separate processes: 329 passed (99 + 62
+ 58 + 29 + 38 + 34 + 9), identical to the final review's count, no flake.
No new logger.* (diagnostic inventory exit 0). No CSS touched.
