+++ Tests/UI/test_library_media_reader_flow.py
@@ -19,6 +19,9 @@ from Tests.UI.test_library_shell import (
     LibraryGlobalKeyProductionCSSHarness,
     LibraryProductionCSSHarness,
     _open_media_find,
+    _painted_cells,
+    _row_is_painted_focused,
+    _top_border_row,
     StaticLibraryMediaScopeService,
     _seed_conversations,
     _two_conversations,
@@ -1929,35 +1932,6 @@ def _focused_id(screen) -> str:
     return str(getattr(screen.focused, "id", "") or "")
 
 
-def _top_border_row(host, box) -> str:
-    """Painted top-border row of one widget, glyphs included (task-31634 cue)."""
-    region = box.region
-    strips = list(host.screen._compositor.render_strips())
-    return strips[region.y].crop(region.x, region.right).text
-
-
-def _painted_cells(host, region):
-    """(text, style) for every non-blank painted segment inside ``region``."""
-    strips = list(host.screen._compositor.render_strips())
-    cells = []
-    for y in range(region.y, min(region.bottom, len(strips))):
-        for segment in strips[y].crop(region.x, region.right):
-            if segment.text.strip():
-                cells.append((segment.text, segment.style))
-    return cells
-
-
-def _row_is_painted_focused(host, row) -> bool:
-    """Whether ``row`` really carries the media row focus cue on screen.
-
-    ``.library-media-row:focus`` sets ``outline: none`` and paints its cue as
-    a STYLE (focus background + bold underline), so a region assertion cannot
-    tell a focused row from an unfocused one (the task-31221 lesson).
-    """
-    cells = _painted_cells(host, row.region)
-    return bool(cells) and all(style.underline for _text, style in cells)
-
-
 async def _settle(pilot, times: int = 4) -> None:
     for _ in range(times):
         await pilot.pause()
@@ -2205,6 +2179,11 @@ async def test_opening_find_costs_no_extra_focus_move(size):
         assert screen._library_media_find_open is True
         assert _focused_id(screen) == "library-media-content-search", screen.focused
         assert "library-media-reader-find" not in moves, moves
-        assert len(moves) <= 2, moves
+        # Exact list, not just a ceiling (review M-5): an inequality would let
+        # a future regression that silently DROPS a legitimate move pass.
+        assert moves == [
+            "library-media-items-grip",
+            "library-media-content-search",
+        ], moves
         for media_id in tuple(service.detail_release):
             service.release(media_id)
